Сканер перенесён в AddTulaActivity.java Лишние активити удалены
