+++ app/src/main/res/layout/activity_add_tula.xml
@@ -8,7 +8,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".presentation.activity.AddTulaManuallyActivity">
+    tools:context=".presentation.activity.AddTulaActivity">
 
     <!-- Название сорта (sort) -->
     <LinearLayout
@@ -74,13 +74,27 @@
             android:text="Добавить"/>
     </LinearLayout>
 
+    <!-- Макет кнопки Сканировать -->
+    <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+        <Button
+                android:id="@+id/bt_scan_tula"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:text="Сканировать"/>
+    </LinearLayout>
+
     <!-- RecyclerView предназначен для оптимизации работы со списками -->
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view_tula_manually"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_marginTop="8dp"
-        tools:listitem="@layout/list_row_tula_manually"/>
+        tools:listitem="@layout/list_row_add_tula"/>
     <!-- В "list_row_main" прописываем элемент списка -->
 
 </LinearLayout>
