+++ app/src/main/res/layout/list_row_add_tula.xml
